Refactor for clearer separation between haproxy-route and haproxy-route-tcp
