+++ charmcraft.yaml
@@ -30,8 +30,10 @@ links:
 requires:
   haproxy-route:
     interface: haproxy-route
+    limit: 1
   haproxy-route-tcp:
     interface: haproxy-route-tcp
+    limit: 1
 
 provides:
   ingress:
@@ -232,6 +234,8 @@ charm-libs:
     version: "1"
   - lib: traefik_k8s.ingress
     version: "2"
+  - lib: gateway_api_integrator.gateway_route
+    version: "1"
 
 actions:
   get-proxied-endpoints:
