+++ pyproject.toml
@@ -27,7 +27,6 @@ fmt = [
 lint = [
   "codespell",
   "jubilant",
-  "lightkube>=0.15.0",
   "mypy",
   "ops[testing]",
   "pep8-naming",
@@ -39,7 +38,6 @@ lint = [
 ]
 unit = [
   "coverage[toml]",
-  "lightkube>=0.15.0",
   "ops[testing]",
   "pytest",
 ]
@@ -54,7 +52,6 @@ integration = [
   "allure-pytest>=2.8.18",
   "allure-pytest-collection-report @ git+https://github.com/canonical/data-platform-workflows@v24.0.0#subdirectory=python/pytest_plugins/allure_pytest_collection_report",
   "jubilant",
-  "lightkube>=0.15.0",
   "pytest",
   "pytest-jubilant",
   "requests",
@@ -64,7 +61,7 @@ integration = [
 package = false
 
 [tool.ruff]
-target-version = "py310"
+target-version = "py312"
 line-length = 99
 
 # enable ruff linters:
