+++ tests/integration/conftest.py
@@ -278,10 +278,10 @@ def application_with_tcp_server_fixture(application: str, juju: jubilant.Juju):
         The ingress-configurator app name.
     """
     juju.wait(
-        lambda status: jubilant.all_active(status, application),
+        lambda status: jubilant.all_agents_idle(status, application),
     )
     command = "sudo snap install ping-pong-tcp; sudo snap set ping-pong-tcp host=0.0.0.0"
-    juju.ssh(target=f"{application}/0", command=command)
+    juju.ssh(target=f"{application}/leader", command=command)
     yield application
