+++ tests/unit/test_kubernetes.py
@@ -15,7 +15,8 @@ from kubernetes import (
     get_nodeport_service,
     get_nodes_ips,
 )
-from state.charm_state import InvalidStateError, NodePortState
+from state.charm_state import InvalidStateError
+from state.haproxy_route import NodePortState
 
 
 def _make_node(*addresses: tuple[str, str]) -> MagicMock:
