+++ tests/unit/test_state_haproxy_route_tcp.py
@@ -10,8 +10,8 @@ from charms.haproxy.v1.haproxy_route_tcp import LoadBalancingAlgorithm
 from ops import CharmBase
 
 from state.haproxy_route_tcp import (
-    HaproxyRouteTcpRequirements,
-    InvalidHaproxyRouteTcpRequirementsError,
+    HaproxyRouteTcpState,
+    InvalidStateError,
 )
 
 
@@ -35,7 +35,7 @@ def test_haproxy_route_tcp_requirements_from_charm():
         "tcp-enforce-tls": True,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert [str(addr) for addr in requirements.backend_addresses] == [
         "192.168.1.10",
@@ -69,8 +69,8 @@ def test_haproxy_route_tcp_invalid_algorithm():
         "tcp-load-balancing-algorithm": "invalid",
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid load balancing algorithm" in str(exc_info.value)
 
 
@@ -92,7 +92,7 @@ def test_haproxy_route_tcp_requirements_from_charm_no_tls():
         "tcp-load-balancing-consistent-hashing": False,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert [str(addr) for addr in requirements.backend_addresses] == ["10.0.0.1"]
     assert requirements.port == 9000
@@ -119,7 +119,7 @@ def test_haproxy_route_tcp_requirements_from_charm_ipv6():
         "tcp-enforce-tls": True,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert [str(addr) for addr in requirements.backend_addresses] == ["2001:db8::1", "2001:db8::2"]
     assert requirements.port == 443
@@ -154,7 +154,7 @@ def test_haproxy_route_tcp_requirements_valid_wildcard_sni(hostname):
         "tcp-enforce-tls": True,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert requirements.hostname == hostname
 
@@ -188,8 +188,8 @@ def test_haproxy_route_tcp_requirements_invalid_wildcard_sni(invalid_hostname):
         "tcp-enforce-tls": True,
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid haproxy-route-tcp configuration" in str(exc_info.value)
 
 
@@ -210,8 +210,8 @@ def test_haproxy_route_tcp_requirements_empty_backend_addresses():
         "tcp-load-balancing-consistent-hashing": False,
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid haproxy-route-tcp configuration" in str(exc_info.value)
 
 
@@ -239,8 +239,8 @@ def test_haproxy_route_tcp_requirements_invalid_backend_address(invalid_address)
         "tcp-hostname": None,
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid load balancing algorithm" in str(exc_info.value)
 
 
@@ -270,8 +270,8 @@ def test_haproxy_route_tcp_requirements_invalid_frontend_port(invalid_port):
         "tcp-load-balancing-consistent-hashing": False,
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid haproxy-route-tcp configuration" in str(exc_info.value)
 
 
@@ -301,8 +301,8 @@ def test_haproxy_route_tcp_requirements_invalid_backend_port(invalid_port):
         "tcp-load-balancing-consistent-hashing": False,
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid haproxy-route-tcp configuration" in str(exc_info.value)
 
 
@@ -324,7 +324,7 @@ def test_haproxy_route_tcp_requirements_valid_port_boundaries():
         "tcp-enforce-tls": True,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert requirements.port == 1
     assert requirements.backend_port == 65535
@@ -348,7 +348,7 @@ def test_haproxy_route_tcp_requirements_multiple_backend_addresses():
         "tcp-enforce-tls": True,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert len(requirements.backend_addresses) == 3
     assert [str(addr) for addr in requirements.backend_addresses] == [
@@ -376,7 +376,7 @@ def test_haproxy_route_tcp_requirements_mixed_ip_versions():
         "tcp-enforce-tls": True,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert len(requirements.backend_addresses) == 2
     assert [str(addr) for addr in requirements.backend_addresses] == [
@@ -412,7 +412,7 @@ def test_haproxy_route_tcp_requirements_enforce_tls(
         "tcp-load-balancing-consistent-hashing": False,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert requirements.enforce_tls is expected_enforce_tls
     assert requirements.hostname == expected_hostname
@@ -456,7 +456,7 @@ def test_haproxy_route_tcp_requirements_health_check_presence(
         **health_check_config,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert requirements.health_check.interval == expected_interval
     assert requirements.health_check.rise == expected_rise
@@ -489,8 +489,8 @@ def test_haproxy_route_tcp_requirements_incomplete_health_check(missing_field):
         **missing_field,
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid haproxy-route-tcp configuration" in str(exc_info.value)
 
 
@@ -520,8 +520,8 @@ def test_haproxy_route_tcp_requirements_invalid_health_check_values(invalid_valu
         **invalid_value,
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid haproxy-route-tcp configuration" in str(exc_info.value)
 
 
@@ -549,7 +549,7 @@ def test_haproxy_route_tcp_requirements_health_check_with_type_generic():
         "tcp-health-check-expect": "PONG",
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert requirements.health_check.check_type is not None
     assert requirements.health_check.check_type.value == "generic"
@@ -581,7 +581,7 @@ def test_haproxy_route_tcp_requirements_health_check_with_type_mysql():
         "tcp-health-check-db-user": "health_checker",
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert requirements.health_check.check_type is not None
     assert requirements.health_check.check_type.value == "mysql"
@@ -616,7 +616,7 @@ def test_haproxy_route_tcp_requirements_valid_health_check_types(check_type):
         "tcp-health-check-type": check_type,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert requirements.health_check.check_type is not None
     assert requirements.health_check.check_type.value == check_type
@@ -643,8 +643,8 @@ def test_haproxy_route_tcp_requirements_invalid_health_check_type():
         "tcp-health-check-type": "invalid_type",
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid health check type" in str(exc_info.value)
 
 
@@ -686,7 +686,7 @@ def test_haproxy_route_tcp_requirements_timeout_presence(
         **timeout_config,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert requirements.timeout.server == expected_server
     assert requirements.timeout.connect == expected_connect
@@ -719,8 +719,8 @@ def test_haproxy_route_tcp_requirements_invalid_timeout_values(invalid_value):
         **invalid_value,
     }
 
-    with pytest.raises(InvalidHaproxyRouteTcpRequirementsError) as exc_info:
-        HaproxyRouteTcpRequirements.from_charm(charm)
+    with pytest.raises(InvalidStateError) as exc_info:
+        HaproxyRouteTcpState.from_charm(charm)
     assert "Invalid haproxy-route-tcp configuration" in str(exc_info.value)
 
 
@@ -743,6 +743,6 @@ def test_haproxy_route_tcp_requirements_proxy_protocol_enabled():
         "tcp-enable-proxy-protocol": True,
     }
 
-    requirements = HaproxyRouteTcpRequirements.from_charm(charm)
+    requirements = HaproxyRouteTcpState.from_charm(charm)
 
     assert requirements.proxy_protocol is True
